requirements.txt
aiogram==3.3.0
supabase==2.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
flask==3.0.0
requests==2.31.0